requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
pymongo>=4.6.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
